+++ packages/cli/src/__tests__/ceramic-daemon.test.ts
@@ -20,7 +20,6 @@ import { randomString } from '@stablelib/random'
 import { CommitID, StreamID } from '@ceramicnetwork/streamid'
 import { createIPFS } from '@ceramicnetwork/ipfs-daemon'
 import { makeDID } from './make-did.js'
-import fetch from 'cross-fetch'
 import { makeCeramicCore } from './make-ceramic-core.js'
 import { makeCeramicDaemon } from './make-ceramic-daemon.js'
 import { DID } from 'dids'
@@ -212,7 +211,7 @@ describe('Ceramic interop: core <> http-client', () => {
     // change from core viewable in client
     await doc1.update(middleContent)
     await anchorDoc(doc1)
-    await TestUtils.waitForConditionOrTimeout(() => {
+    await TestUtils.waitForConditionOrTimeout(async () => {
       return JSON.stringify(doc1.content) === JSON.stringify(doc2.content)
     })
     expect(doc1.content).toEqual(middleContent)
@@ -222,7 +221,7 @@ describe('Ceramic interop: core <> http-client', () => {
 
     await doc2.update(finalContent)
     await anchorDoc(doc2)
-    await TestUtils.waitForConditionOrTimeout(() => {
+    await TestUtils.waitForConditionOrTimeout(async () => {
       return JSON.stringify(doc1.content) === JSON.stringify(doc2.content)
     })
     expect(doc1.content).toEqual(doc2.content)