+++ packages/cli/src/ceramic-daemon.ts
@@ -13,17 +13,15 @@ import {
   LogStyle,
   ModelData,
   MultiQuery,
+  Networks,
   StreamUtils,
   SyncOptions,
   UnreachableCaseError,
 } from '@ceramicnetwork/common'
-import { StreamID, StreamType } from '@ceramicnetwork/streamid'
+import { StreamID } from '@ceramicnetwork/streamid'
 import * as ThreeIdResolver from '@ceramicnetwork/3id-did-resolver'
 import * as KeyDidResolver from 'key-did-resolver'
 import * as PkhDidResolver from 'pkh-did-resolver'
-import EthrDidResolver from 'ethr-did-resolver'
-import * as NftDidResolver from 'nft-did-resolver'
-import * as SafeDidResolver from 'safe-did-resolver'
 
 import { DID, DIDOptions, DIDProvider } from 'dids'
 import cors from 'cors'
@@ -32,7 +30,7 @@ import { addAsync, ExpressWithAsync } from '@awaitjs/express'
 import { instrumentRequests } from './daemon/instrument-requests.js'
 import { logRequests } from './daemon/log-requests.js'
 import type { Server } from 'http'
-import { DaemonConfig, StateStoreMode } from './daemon-config.js'
+import { DaemonConfig, IpfsMode, StateStoreMode } from './daemon-config.js'
 import type { ResolverRegistry } from 'did-resolver'
 import { ErrorHandlingRouter } from './error-handling-router.js'
 import { collectionQuery, countQuery } from './daemon/collection-queries.js'
@@ -58,12 +56,8 @@ const STREAM_UNPINNED = 'stream_unpinned'
 type AdminCode = string
 type Timestamp = number
 
-interface MultiQueryWithDocId extends MultiQuery {
-  docId?: string
-}
-
 interface MultiQueries {
-  queries: Array<MultiQueryWithDocId>
+  queries: Array<MultiQuery>
   timeout?: number
 }
 
@@ -107,7 +101,7 @@ export function makeCeramicConfig(opts: DaemonConfig): CeramicConfig {
 
   const ceramicConfig: CeramicConfig = {
     loggerProvider,
-    gateway: opts.node.gateway || false,
+    readOnly: opts.node.readOnly || false,
     anchorServiceUrl: opts.anchor.anchorServiceUrl,
     anchorServiceAuthMethod: opts.anchor.authMethod,
     ethereumRpcUrl: opts.anchor.ethereumRpcUrl,
@@ -173,44 +167,12 @@ function upconvertLegacySyncOption(opts: Record<string, any> | undefined) {
 /**
  * Prepare DID resolvers to use in the daemon.
  */
-function makeResolvers(
-  ceramic: Ceramic,
-  ceramicConfig: CeramicConfig,
-  opts: DaemonConfig
-): ResolverRegistry {
-  let result = {
+function makeResolvers(ceramic: Ceramic): ResolverRegistry {
+  return {
     ...KeyDidResolver.getResolver(),
     ...PkhDidResolver.getResolver(),
     ...ThreeIdResolver.getResolver(ceramic),
-    ...NftDidResolver.getResolver({
-      ceramic: ceramic,
-      ...opts.didResolvers?.nftDidResolver,
-    }),
-    ...SafeDidResolver.getResolver({
-      ceramic: ceramic,
-      ...opts.didResolvers?.safeDidResolver,
-    }),
   }
-  if (
-    opts.didResolvers?.ethrDidResolver?.networks &&
-    opts.didResolvers?.ethrDidResolver?.networks.length > 0
-  ) {
-    // Custom ethr-did-resolver configuration passed
-    result = { ...result, ...EthrDidResolver.getResolver(opts.didResolvers.ethrDidResolver) }
-  } else if (ceramicConfig.ethereumRpcUrl) {
-    // Use default network from ceramic config's ethereumRpcUrl
-    result = {
-      ...result,
-      ...EthrDidResolver.getResolver({
-        networks: [
-          {
-            rpcUrl: ceramicConfig.ethereumRpcUrl,
-          },
-        ],
-      }),
-    }
-  }
-  return result
 }
 
 /**
@@ -326,7 +288,7 @@ export class CeramicDaemon {
 
     this.app.use(logRequests(ceramic.loggerProvider))
 
-    this.registerAPIPaths(this.app, opts.node?.gateway)
+    this.registerAPIPaths(this.app, opts.node?.readOnly)
 
     this.app.use(errorHandler(this.diagnosticsLogger))
   }
@@ -348,6 +310,11 @@ export class CeramicDaemon {
   static async create(opts: DaemonConfig): Promise<CeramicDaemon> {
     const ceramicConfig = makeCeramicConfig(opts)
 
+    if (opts.ipfs.mode != IpfsMode.REMOTE && opts.network?.name == Networks.MAINNET) {
+      throw new Error(
+        "Cannot run IPFS in 'bundled' mode on Mainnet. Bundled IPFS is for testing only and not suitable for production deployments.  Please set up a dedicated IPFS node and provide the URL for a REMOTE connection"
+      )
+    }
     const ipfs = await IpfsConnectionFactory.buildIpfsConnection(
       opts.ipfs.mode,
       opts.network?.name,
@@ -380,7 +347,7 @@ export class CeramicDaemon {
       await ceramic.repository.injectKeyValueStore(s3Store)
     }
 
-    let didOptions: DIDOptions = { resolver: makeResolvers(ceramic, ceramicConfig, opts) }
+    let didOptions: DIDOptions = { resolver: makeResolvers(ceramic) }
     let provider: DIDProvider
 
     if (opts.node.sensitive_privateSeedUrl()) {
@@ -411,13 +378,11 @@ export class CeramicDaemon {
     return daemon
   }
 
-  registerAPIPaths(app: ExpressWithAsync, gateway: boolean): void {
+  registerAPIPaths(app: ExpressWithAsync, readOnly: boolean): void {
     const baseRouter = ErrorHandlingRouter(this.diagnosticsLogger)
     const commitsRouter = ErrorHandlingRouter(this.diagnosticsLogger)
-    const documentsRouter = ErrorHandlingRouter(this.diagnosticsLogger)
     const multiqueriesRouter = ErrorHandlingRouter(this.diagnosticsLogger)
     const nodeRouter = ErrorHandlingRouter(this.diagnosticsLogger)
-    const recordsRouter = ErrorHandlingRouter(this.diagnosticsLogger)
     const streamsRouter = ErrorHandlingRouter(this.diagnosticsLogger)
     const collectionRouter = ErrorHandlingRouter(this.diagnosticsLogger)
     const adminCodesRouter = ErrorHandlingRouter(this.diagnosticsLogger)
@@ -430,10 +395,8 @@ export class CeramicDaemon {
 
     app.use('/api/v0', baseRouter)
     baseRouter.use('/commits', commitsRouter)
-    baseRouter.use('/documents', documentsRouter)
     baseRouter.use('/multiqueries', multiqueriesRouter)
     baseRouter.use('/node', nodeRouter)
-    baseRouter.use('/records', recordsRouter)
     baseRouter.use('/streams', streamsRouter)
     baseRouter.use('/collection', collectionRouter)
     baseRouter.use('/admin/getCode', adminCodesRouter)
@@ -459,11 +422,7 @@ export class CeramicDaemon {
     streamsRouter.getAsync('/:streamid/content', this.content.bind(this))
     nodeRouter.getAsync('/chains', this.getSupportedChains.bind(this))
     nodeRouter.getAsync('/healthcheck', this.healthcheck.bind(this))
-    documentsRouter.getAsync('/:docid', this.stateOld.bind(this)) // Deprecated
-    recordsRouter.getAsync('/:streamid', this.commits.bind(this)) // Deprecated
-    collectionRouter.getAsync('/', this.getCollection_get.bind(this)) // Deprecated
     collectionRouter.postAsync('/', this.getCollection_post.bind(this))
-    collectionRouter.getAsync('/count', this.getCollectionCount_get.bind(this)) // Deprecated
     collectionRouter.postAsync('/count', this.getCollectionCount_post.bind(this))
     adminCodesRouter.getAsync('/', this.getAdminCode.bind(this))
     adminNodeStatusRouter.getAsync('/', this.nodeStatus.bind(this))
@@ -477,21 +436,15 @@ export class CeramicDaemon {
     adminPinsRouter.getAsync('/', this.listPinned.bind(this))
     adminShutdownRouter.postAsync('/', this.shutdownServer.bind(this))
 
-    if (!gateway) {
+    if (!readOnly) {
       streamsRouter.postAsync('/', this.createStreamFromGenesis.bind(this))
       streamsRouter.postAsync('/:streamid/anchor', this.requestAnchor.bind(this))
       commitsRouter.postAsync('/', this.applyCommit.bind(this))
       adminPinsRouter.postAsync('/:streamid', this.pinStream.bind(this))
       adminPinsRouter.deleteAsync('/:streamid', this.unpinStream.bind(this))
-
-      documentsRouter.postAsync('/', this.createDocFromGenesis.bind(this)) // Deprecated
-      recordsRouter.postAsync('/', this.applyCommit.bind(this)) // Deprecated
     } else {
       streamsRouter.postAsync('/', this.createReadOnlyStreamFromGenesis.bind(this))
       commitsRouter.postAsync('/', this._notSupported.bind(this))
-
-      documentsRouter.postAsync('/', this.createReadOnlyDocFromGenesis.bind(this)) // Deprecated
-      recordsRouter.postAsync('/', this._notSupported.bind(this)) // Deprecated
     }
   }
 
@@ -564,28 +517,7 @@ export class CeramicDaemon {
   }
 
   /**
-   * Create document from genesis commit
-   * @dev Useful when the streamId is unknown, but you have the genesis contents
-   * @deprecated
-   */
-  async createDocFromGenesis(req: Request, res: Response): Promise<void> {
-    const { doctype, genesis, docOpts } = req.body
-    upconvertLegacySyncOption(docOpts)
-    const type = StreamType.codeByName(doctype)
-    const doc = await this.ceramic.createStreamFromGenesis(
-      type,
-      StreamUtils.deserializeCommit(genesis),
-      docOpts
-    )
-    res.json({
-      streamId: doc.id.toString(),
-      docId: doc.id.toString(),
-      state: StreamUtils.serializeState(doc.state),
-    })
-  }
-
-  /**
-   * Create document from genesis commit
+   * Create stream from genesis commit
    * @dev Useful when the streamId is unknown, but you have the genesis contents
    */
   async createStreamFromGenesis(req: Request, res: Response): Promise<void> {
@@ -609,34 +541,9 @@ export class CeramicDaemon {
   }
 
   /**
-   * Create read-only document from genesis commit
+   * Create read-only stream from genesis commit
    * @dev Useful when the streamId is unknown, but you have the genesis contents
-   * @TODO Should return null if document does not already exist instead of
-   * current behavior, publishing to IPFS. With that change it will make sense
-   * to rename this, e.g. `loadStreamFromGenesis`
-   * @deprecated
-   */
-  async createReadOnlyDocFromGenesis(req: Request, res: Response): Promise<void> {
-    const { doctype, genesis, docOpts } = req.body
-    upconvertLegacySyncOption(docOpts)
-    const type = StreamType.codeByName(doctype)
-    const readOnlyDocOpts = { ...docOpts, anchor: false, publish: false }
-    const doc = await this.ceramic.createStreamFromGenesis(
-      type,
-      StreamUtils.deserializeCommit(genesis),
-      readOnlyDocOpts
-    )
-    res.json({
-      streamId: doc.id.toString(),
-      docId: doc.id.toString(),
-      state: StreamUtils.serializeState(doc.state),
-    })
-  }
-
-  /**
-   * Create read-only document from genesis commit
-   * @dev Useful when the docId is unknown, but you have the genesis contents
-   * @TODO Should return null if document does not already exist instead of
+   * @TODO Should return null if stream does not already exist instead of
    * current behavior, publishing to IPFS. With that change it will make sense
    * to rename this, e.g. `loadStreamFromGenesis`
    */
@@ -661,22 +568,10 @@ export class CeramicDaemon {
   }
 
   /**
-   * Get document state
-   * @deprecated
-   * // todo remove when 'documents' endpoint is removed
-   */
-  async stateOld(req: Request, res: Response): Promise<void> {
-    const opts = parseQueryObject(req.query)
-    upconvertLegacySyncOption(opts)
-    const doc = await this.ceramic.loadStream(req.params.docid, opts)
-    res.json({ docId: doc.id.toString(), state: StreamUtils.serializeState(doc.state) })
-  }
-
-  /**
-   * Get all document commits
+   * Get all stream commits
    */
   async commits(req: Request, res: Response): Promise<void> {
-    const streamId = StreamID.fromString(req.params.streamid || req.params.docid)
+    const streamId = StreamID.fromString(req.params.streamid)
     const commits = await this.ceramic.loadStreamCommits(streamId)
     const serializedCommits = commits.map((r: any) => {
       return {
@@ -685,23 +580,10 @@ export class CeramicDaemon {
       }
     })
 
-    // TODO remove docId from output when we are no longer supporting clients older than v1.0.0
     res.json({
       streamId: streamId.toString(),
-      docId: streamId.toString(),
       commits: serializedCommits,
     })
-  }
-
-  /**
-   * Implementation of collection queries (the 'collection' http endpoint) using the HTTP GET method.
-   * TODO: Remove this once we no longer need to support version 2.4.0 and earlier of the
-   * http-client.
-   */
-  async getCollection_get(req: Request, res: Response): Promise<void> {
-    const httpQuery = parseQueryObject(req.query)
-    const response = await this._getCollection(httpQuery)
-    res.json(response)
   }
 
   /**
@@ -729,17 +611,6 @@ export class CeramicDaemon {
       }),
       pageInfo: indexResponse.pageInfo,
     }
-  }
-
-  /**
-   * Implementation of count queries (the 'collection/count' http endpoint) using the HTTP GET method.
-   * TODO: Remove this once we no longer need to support version 2.4.0 and earlier of the
-   * http-client.
-   */
-  async getCollectionCount_get(req: Request, res: Response): Promise<void> {
-    const httpQuery = parseQueryObject(req.query)
-    const response = await this._getCollectionCount(httpQuery)
-    res.json(response)
   }
 
   /**
@@ -1040,14 +911,14 @@ export class CeramicDaemon {
   }
 
   /**
-   * Apply one commit to the existing document
+   * Apply one commit to the existing stream
    */
   async applyCommit(req: Request, res: Response): Promise<void> {
-    const { docId, commit, docOpts } = req.body
-    const opts = req.body.opts || docOpts
+    const { commit } = req.body
+    const opts = req.body.opts
     upconvertLegacySyncOption(opts)
 
-    const streamId = req.body.streamId || docId
+    const streamId = req.body.streamId
     if (!(streamId && commit)) {
       throw new Error('streamId and commit are required in order to apply commit')
     }
@@ -1059,27 +930,16 @@ export class CeramicDaemon {
     )
     res.json({
       streamId: stream.id.toString(),
-      docId: stream.id.toString(),
       state: StreamUtils.serializeState(stream.state),
     })
   }
 
   /**
-   * Load multiple documents and paths using an array of multiqueries
+   * Load multiple streams and paths using an array of multiqueries
    */
   async multiQuery(req: Request, res: Response): Promise<void> {
-    let { queries } = <MultiQueries>req.body
+    const { queries } = <MultiQueries>req.body
     const { timeout } = <MultiQueries>req.body
-
-    // Handle queries from old clients by replacing the `docId` arguments with `streamId`.
-    // TODO: Remove this once we no longer need to support http clients older than version 1.0.0
-    queries = queries.map((q: MultiQueryWithDocId): MultiQuery => {
-      if (q.docId) {
-        q.streamId = q.docId
-        delete q.docId
-      }
-      return q
-    })
 
     const results = await this.ceramic.multiQuery(queries, timeout)
     const response = Object.entries(results).reduce((acc, e) => {
@@ -1103,13 +963,12 @@ export class CeramicDaemon {
    * Pin stream
    */
   async pinStream(req: Request, res: Response): Promise<void> {
-    const streamId = StreamID.fromString(req.params.streamid || req.params.docid)
+    const streamId = StreamID.fromString(req.params.streamid)
     const { force } = req.body
     await this.ceramic.admin.pin.add(streamId, force)
     Metrics.count(STREAM_PINNED, 1)
     res.json({
       streamId: streamId.toString(),
-      docId: streamId.toString(),
       isPinned: true,
     })
   }
@@ -1118,13 +977,12 @@ export class CeramicDaemon {
    * Unpin stream
    */
   async unpinStream(req: Request, res: Response): Promise<void> {
-    const streamId = StreamID.fromString(req.params.streamid || req.params.docid)
+    const streamId = StreamID.fromString(req.params.streamid)
     const { opts } = req.body
     await this.ceramic.admin.pin.rm(streamId, opts)
     Metrics.count(STREAM_UNPINNED, 1)
     res.json({
       streamId: streamId.toString(),
-      docId: streamId.toString(),
       isPinned: false,
     })
   }
@@ -1134,8 +992,8 @@ export class CeramicDaemon {
    */
   async listPinned(req: Request, res: Response): Promise<void> {
     let streamId: StreamID
-    if (req.params.streamid || req.params.docid) {
-      streamId = StreamID.fromString(req.params.streamid || req.params.docid)
+    if (req.params.streamid) {
+      streamId = StreamID.fromString(req.params.streamid)
     }
     const pinnedStreamIds = []
     const iterator = await this.ceramic.admin.pin.ls(streamId)
@@ -1143,16 +1001,11 @@ export class CeramicDaemon {
       pinnedStreamIds.push(id)
     }
 
-    // Return the same data in two formats: 'pinnedStreamids' and 'pinnedDocIds', to support old clients
-    // TODO: Remove 'pinnedDocIds' once we are okay with no longer supporting applications using a
-    // version of '@ceramicnetwork/http-client' older than v1.0.0
-    res.json({ pinnedStreamIds, pinnedDocIds: pinnedStreamIds })
+    res.json({ pinnedStreamIds })
   }
 
   async _notSupported(req: Request, res: Response): Promise<void> {
-    res
-      .status(StatusCodes.BAD_REQUEST)
-      .json({ error: 'Method not supported by read only Ceramic Gateway' })
+    res.status(StatusCodes.BAD_REQUEST).json({ error: 'Method not supported by read only Ceramic' })
   }
 
   async _pinNotSupported(req: Request, res: Response): Promise<void> {