+++ packages/core/src/store/__tests__/level-state-store.test.ts
@@ -9,7 +9,7 @@ import {
   Networks,
   DiagnosticsLogger,
 } from '@ceramicnetwork/common'
-import { LevelDbStore } from '../level-db-store.js'
+import { LevelDbStore, OLD_ELP_DEFAULT_LOCATION } from '../level-db-store.js'
 import { StreamStateStore } from '../stream-state-store.js'
 
 class FakeType extends Stream {
@@ -170,7 +170,7 @@ describe('LevelDB-backed StateStore network change tests', () => {
   })
 
   test('switch from ELP to Mainnet preserves data', async () => {
-    const elpLevelStore = new LevelDbStore(logger, tmpFolder.path, Networks.ELP)
+    const elpLevelStore = new LevelDbStore(logger, tmpFolder.path, OLD_ELP_DEFAULT_LOCATION)
     await stateStore.open(elpLevelStore)
 
     const state = TestUtils.makeStreamState()