+++ packages/indexing/src/__tests__/insertion-order.test.ts
@@ -12,7 +12,7 @@ import tmp from 'tmp-promise'
 
 const MODEL_ID = 'kjzl6cwe1jw145m7jxh4jpa6iw1ps3jcjordpo81e0w04krcpz8knxvg5ygiabd'
 const MODELS_TO_INDEX = [StreamID.fromString(MODEL_ID)]
-const MODEL = MODELS_TO_INDEX[0]
+const MODELS = new Set([MODEL_ID])
 const logger = new LoggerProvider().getDiagnosticsLogger()
 
 let EXPECTED: Array<string>
@@ -67,8 +67,7 @@ describe('postgres', () => {
       const pages = chunks(EXPECTED, pageSize)
       let afterCursor: string | undefined = undefined
       for (let i = 0; i < pages.length; i++) {
-        const result = await order.page({
-          model: MODEL,
+        const result = await order.page(MODELS, {
           first: pageSize,
           after: afterCursor,
         })
@@ -87,14 +86,12 @@ describe('postgres', () => {
       const pageSize = 5
       const pages = chunks(EXPECTED, pageSize)
       const expectedFirstPage = pages[0]
-      const firstPage = await order.page({
-        model: MODEL,
+      const firstPage = await order.page(MODELS, {
         first: pageSize,
       })
       expect(firstPage.edges.map((e) => String(e.node))).toEqual(expectedFirstPage)
       const secondEntry = firstPage.edges[1]
-      const customPage = await order.page({
-        model: MODEL,
+      const customPage = await order.page(MODELS, {
         first: 3,
         after: secondEntry.cursor,
       })
@@ -114,8 +111,7 @@ describe('postgres', () => {
     test('pagination', async () => {
       let beforeCursor: string | undefined = undefined
       for (let i = 0; i < pages.length; i++) {
-        const result = await order.page({
-          model: MODEL,
+        const result = await order.page(MODELS, {
           last: PAGE_SIZE,
           before: beforeCursor,
         })
@@ -133,14 +129,12 @@ describe('postgres', () => {
 
     test('using edge cursor', async () => {
       const expectedFirstPage = pages[0]
-      const firstPage = await order.page({
-        model: MODEL,
+      const firstPage = await order.page(MODELS, {
         last: PAGE_SIZE,
       })
       expect(firstPage.edges.map((e) => String(e.node))).toEqual(expectedFirstPage)
       const secondLastEntry = firstPage.edges[firstPage.edges.length - 2]
-      const customPage = await order.page({
-        model: MODEL,
+      const customPage = await order.page(MODELS, {
         last: 3,
         before: secondLastEntry.cursor,
       })
@@ -153,15 +147,13 @@ describe('postgres', () => {
   test('filtered by account', async () => {
     const presentAccount = 'did:key:foo' // We have that in the populated table
     const absentAccount = 'did:key:absent' // We do not have that in the populated table
-    const withPresentAccount = await order.page({
-      model: MODEL,
+    const withPresentAccount = await order.page(MODELS, {
       account: presentAccount,
       first: 5,
     })
     expect(withPresentAccount.edges.map((e) => String(e.node))).toEqual(EXPECTED.slice(0, 5))
     // Should return an empty page
-    const withAbsentAccount = await order.page({
-      model: MODEL,
+    const withAbsentAccount = await order.page(MODELS, {
       account: absentAccount,
       first: 5,
     })
@@ -214,8 +206,7 @@ describe('sqlite', () => {
       const pages = chunks(EXPECTED, pageSize)
       let afterCursor: string | undefined = undefined
       for (let i = 0; i < pages.length; i++) {
-        const result = await order.page({
-          model: MODEL,
+        const result = await order.page(MODELS, {
           first: pageSize,
           after: afterCursor,
         })
@@ -234,14 +225,12 @@ describe('sqlite', () => {
       const pageSize = 5
       const pages = chunks(EXPECTED, pageSize)
       const expectedFirstPage = pages[0]
-      const firstPage = await order.page({
-        model: MODEL,
+      const firstPage = await order.page(MODELS, {
         first: pageSize,
       })
       expect(firstPage.edges.map((e) => String(e.node))).toEqual(expectedFirstPage)
       const secondEntry = firstPage.edges[1]
-      const customPage = await order.page({
-        model: MODEL,
+      const customPage = await order.page(MODELS, {
         first: 3,
         after: secondEntry.cursor,
       })
@@ -261,8 +250,7 @@ describe('sqlite', () => {
     test('pagination', async () => {
       let beforeCursor: string | undefined = undefined
       for (let i = 0; i < pages.length; i++) {
-        const result = await order.page({
-          model: MODEL,
+        const result = await order.page(MODELS, {
           last: PAGE_SIZE,
           before: beforeCursor,
         })
@@ -280,14 +268,12 @@ describe('sqlite', () => {
 
     test('using edge cursor', async () => {
       const expectedFirstPage = pages[0]
-      const firstPage = await order.page({
-        model: MODEL,
+      const firstPage = await order.page(MODELS, {
         last: PAGE_SIZE,
       })
       expect(firstPage.edges.map((e) => String(e.node))).toEqual(expectedFirstPage)
       const secondLastEntry = firstPage.edges[firstPage.edges.length - 2]
-      const customPage = await order.page({
-        model: MODEL,
+      const customPage = await order.page(MODELS, {
         last: 3,
         before: secondLastEntry.cursor,
       })
@@ -300,15 +286,13 @@ describe('sqlite', () => {
   test('filtered by account', async () => {
     const presentAccount = 'did:key:foo' // We have that in the populated table
     const absentAccount = 'did:key:absent' // We do not have that in the populated table
-    const withPresentAccount = await order.page({
-      model: MODEL,
+    const withPresentAccount = await order.page(MODELS, {
       account: presentAccount,
       first: 5,
     })
     expect(withPresentAccount.edges.map((e) => String(e.node))).toEqual(EXPECTED.slice(0, 5))
     // Should return an empty page
-    const withAbsentAccount = await order.page({
-      model: MODEL,
+    const withAbsentAccount = await order.page(MODELS, {
       account: absentAccount,
       first: 5,
     })