+++ packages/ipfs-topology/src/ipfs-topology.ts
@@ -8,7 +8,6 @@ import { Multiaddr, multiaddr } from '@multiformats/multiaddr'
 const BOOTSTRAP_LIST = (ceramicNetwork: Networks): Array<Multiaddr> | null => {
   switch (ceramicNetwork) {
     case Networks.MAINNET:
-    case Networks.ELP:
       return [
         multiaddr(
           '/dns4/go-ipfs-ceramic-private-mainnet-external.3boxlabs.com/tcp/4011/ws/p2p/QmXALVsXZwPWTUbsT8G6VVzzgTJaAWRUD7FWL5f7d5ubAL'
@@ -68,14 +67,19 @@ export class IpfsTopology {
   ) {}
 
   async forceConnection(): Promise<void> {
-    this.logger.debug(`Performing periodic reconnection to bootstrap peers`)
     const bootstrapList: Multiaddr[] = BOOTSTRAP_LIST(this.ceramicNetwork as Networks) || []
     await this._forceBootstrapConnection(this.ipfs, bootstrapList)
   }
 
   async start() {
+    this.logger.imp(`Connecting to bootstrap peers for network ${this.ceramicNetwork}`)
+
     await this.forceConnection()
+    const connectedPeers = (await this.ipfs.swarm.peers()).map((peer) => peer.addr.toString())
+    this.logger.debug(`Connected to peers: ${connectedPeers.join(',')}`)
+
     this.intervalId = setInterval(async () => {
+      this.logger.debug(`Performing periodic reconnection to bootstrap peers for network ${this.ceramicNetwork}`)
       await this.forceConnection()
     }, this.period)
   }