+++ packages/pinning-crust-backend/src/index.ts
@@ -3,7 +3,6 @@ import type { CidList, PinningBackend, PinningInfo } from '@ceramicnetwork/commo
 import { Keyring } from '@polkadot/keyring'
 import { ApiPromise, WsProvider } from '@polkadot/api'
 import { typesBundleForPolkadot } from '@crustio/type-definitions'
-import fetch from 'cross-fetch'
 import * as sha256 from '@stablelib/sha256'
 import { toString } from 'uint8arrays/to-string'
 import { KeyringPair } from '@polkadot/keyring/types'