+++ packages/stream-tests/src/__tests__/deterministic-model-instance-document.test.ts
@@ -1,7 +1,7 @@
 import { jest } from '@jest/globals'
 import getPort from 'get-port'
 import { AnchorStatus, CommitType, IpfsApi, TestUtils } from '@ceramicnetwork/common'
-import { createIPFS } from '@ceramicnetwork/ipfs-daemon'
+import { createIPFS, swarmConnect } from '@ceramicnetwork/ipfs-daemon'
 import {
   ModelInstanceDocument,
   ModelInstanceDocumentMetadataArgs,
@@ -145,6 +145,7 @@ describe('ModelInstanceDocument API multi-node tests', () => {
   beforeAll(async () => {
     ipfs0 = await createIPFS()
     ipfs1 = await createIPFS()
+    await swarmConnect(ipfs0, ipfs1)
   }, 12000)
 
   beforeEach(async () => {