+++ packages/indexing/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.5.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/indexing@1.5.0-rc.1...@ceramicnetwork/indexing@1.5.0) (2023-11-08)
+
+**Note:** Version bump only for package @ceramicnetwork/indexing
+
+
+
+
+
+# [1.5.0-rc.1](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/indexing@1.5.0-rc.0...@ceramicnetwork/indexing@1.5.0-rc.1) (2023-11-06)
+
+**Note:** Version bump only for package @ceramicnetwork/indexing
+
+
+
+
+
+# [1.5.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/indexing@1.4.0...@ceramicnetwork/indexing@1.5.0-rc.0) (2023-11-02)
+
+
+### Features
+
+* **cli,common,indexing:** support interfaces queries ([#3011](https://github.com/ceramicnetwork/js-ceramic/issues/3011)) ([620babd](https://github.com/ceramicnetwork/js-ceramic/commit/620babde2ca913f4571a44a529270b893a1c6668))
+
+
+
+
+
+# [1.4.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/indexing@1.4.0-rc.1...@ceramicnetwork/indexing@1.4.0) (2023-10-26)
+
+**Note:** Version bump only for package @ceramicnetwork/indexing
+
+
+
+
+
 # [1.4.0-rc.1](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/indexing@1.4.0-rc.0...@ceramicnetwork/indexing@1.4.0-rc.1) (2023-10-20)